8-Refactoring EmployeeController

+++ TestNinja/Mocking/EmployeeStorage.cs
@@ -19,8 +19,11 @@ namespace TestNinja.Mocking
         public void DeleteEmployee(int id)
         {
             var employee = _db.Employees.Find(id);
-            _db.Employees.Remove(employee);
-            _db.SaveChanges();
+            if (employee != null)
+            {
+                _db.Employees.Remove(employee);
+                _db.SaveChanges();
+            }
 
 
         }
